rename localsession -> newlocalsession
