--- modules/ui/src/components/LocalSession/index.ts
@@ -1,1 +1,0 @@
-export * from "./LocalSession";
