+++ modules/ui/src/components/NewLocalSession/NewLocalSession.VariantSelector.tsx
