+++ modules/ui/src/components/NewLocalSession/NewLocalSession.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from "react";
-import css from "./LocalSession.cssProxy";
+import css from "./NewLocalSession.cssProxy";
 import {
   AlgolLocalBattle,
   AlgolGameGraphics,
@@ -11,9 +11,9 @@ import {
 import { Button } from "../Button";
 import { SessionList } from "../SessionList";
 import { ImportBattle } from "../ImportBattle";
-import { VariantSelector } from "./LocalSession.VariantSelector";
+import { VariantSelector } from "./NewLocalSession.VariantSelector";
 
-export interface LocalSessionActions {
+export interface NewLocalSessionActions {
   newLocalBattle: (code: string) => void;
   loadLocalSession: (session: AlgolLocalBattle) => void;
   importSession: (str: string) => void;
@@ -21,20 +21,20 @@ export interface LocalSessionActions {
   reportError: AlgolErrorReporter;
 }
 
-type LocalSessionProps = {
+type NewLocalSessionProps = {
   graphics: AlgolGameGraphics;
-  actions: LocalSessionActions;
+  actions: NewLocalSessionActions;
   hasPrevious: boolean;
   meta: AlgolMeta<AlgolGameBlobAnon>;
   variants: AlgolVariantAnon[];
 };
 
-export const LocalSession: FunctionComponent<LocalSessionProps> = props => {
+export const NewLocalSession: FunctionComponent<NewLocalSessionProps> = props => {
   const { actions, meta, graphics, hasPrevious, variants } = props;
   return (
-    <div className={css.localSession}>
+    <div className={css.newLocalSession}>
       <VariantSelector variants={variants} actions={actions} />
-      <div className={css.localSessionDivider} />
+      <div className={css.newLocalSessionDivider} />
       <Button
         disabled={!hasPrevious && "No previous battle found for this game."}
         onClick={actions.continuePreviousSession}
@@ -43,13 +43,13 @@ export const LocalSession: FunctionComponent<LocalSessionProps> = props => {
       >
         Load last battle
       </Button>
-      <div className={css.localSessionDivider} />
+      <div className={css.newLocalSessionDivider} />
       <Button disabled="AI is in the works, but remote play will be implemented first.">
         Versus AI
       </Button>
-      <div className={css.localSessionDivider} />
+      <div className={css.newLocalSessionDivider} />
       <ImportBattle actions={actions} />
-      <div className={css.localSessionDivider} />
+      <div className={css.newLocalSessionDivider} />
       <SessionList
         meta={meta}
         graphics={graphics}
